pyproject.toml
[tool.poetry]
name = "master-server"
version = "0.1.0"
description = "Master server for oxyproxy.io"
authors = ["avilabss <contact@avilabs.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
sendgrid = "^6.11.0"
uvicorn = "^0.30.1"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.7"
pytest = "^8.2.2"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"
aiosqlite = "^0.20.0"
pytest-anyio = "^0.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
